[FIX] mail: Fix noupdate records migration
* Missing security rules noupdate should be removed
* Group > Channel XML-ID should be renamed also for avoiding a duplicate

+++ addons/mail/migrations/9.0.1.0/pre-migration.py
@@ -25,6 +25,10 @@ column_renames = {
     ]
 }
 
+xmlid_renames = [
+    ('mail.group_all_employees', 'mail.channel_all_employees'),
+]
+
 
 @openupgrade.migrate(use_env=True)
 def migrate(env, version):
@@ -32,6 +36,15 @@ def migrate(env, version):
     openupgrade.rename_tables(env.cr, table_renames)
     openupgrade.rename_fields(env, field_renames)
     openupgrade.rename_columns(env.cr, column_renames)
-    # Remove noupdate ir.rule
-    rule = env.ref('mail.mail_group_public_and_joined')
-    rule.unlink()
+    openupgrade.rename_xmlids(env.cr, xmlid_renames)
+    # Remove noupdate records
+    xml_ids = [
+        'mail.mail_group_public_and_joined',
+        'mail_followers_read_write_others',
+        'mail_notification_read_write_own',
+        'ir_cron_mail_garbage_collect_attachments',
+    ]
+    for xml_id in xml_ids:
+        rule = env.ref('mail.' + xml_id, False)
+        if rule:
+            rule.unlink()
